Remove operations create logic from RpcClient class and put it to Wallet class
